python-cycler: Add new recipe for python-cycler package
When using matplotlib and plotting more than one line, it is common to
want to be able to cycle over one or more artist styles; but the
plotting logic can quickly become very involved.  To address this and
allow easy cycling over arbitrary 'kwargs' the `Cycler` class, a
composable kwarg iterator, was developed.

Add recipe to build the python cycler package for both python2 and
python3

Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>