python3-ipykernel: Modify setup.py to pass the correct python executable path
Python3-ipykernel on target uses /usr/share/kernel.json file to specify
arguments that are expanded and run on target. One of the arguments in
kernel.json is the python executable path.

In the 5.3.1 version of python3-ipykernel, the executable path is determined in
setup.py based on how the package is being installed.  If the package is
installed without re-building i.e bdist_egg or wheel packages, then python path
is set to default path.  If the package is built from source then python path
is set to location of sys.executable of python.

This check in bitbake system is an issue as all packages are built from source
and compiled for target.  Since the package is being compiled on host machine,
kernel.json sets the python executable path to PYTHON_PATH of the host machine.
When the package is run on target, it fails with file not found error as it
looks for the PYTHON_PATH on target.

Fix this issue by modifying the setup.py to set the python executable path as
default path so that kernel.json generates the correct command for target.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>