bqplot: Add bqplot and its dependencies
bqplot is a 2-D visualization system for Jupyter, based on the constructs of the
Grammar of Graphics. In bqplot, every component of a plot is an interactive
widget. This allows the user to integrate visualizations with other Jupyter
interactive widgets to create integrated GUIs with a few simple lines of Python
code.

Also include missing dependencies: traittypes.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>